+++ packages/kit/src/runtime/server/utils.js
@@ -137,7 +137,7 @@ export function data_response(data) {
 }
 
 /**
- * @template {'hydrate' | 'prerender' | 'router' | 'ssr'} Option
+ * @template {'prerender' | 'ssr' | 'csr'} Option
  * @template {Option extends 'prerender' ? import('types').PrerenderOption : boolean} Value
  *
  * @param {Array<import('types').SSRNode | undefined>} nodes
@@ -146,13 +146,20 @@ export function data_response(data) {
  * @returns {Value | undefined}
  */
 export function get_option(nodes, option) {
-	return nodes.reduce(
-		(value, node) =>
-			/** @type {any} TypeScript's too dumb to understand this */ (
-				node?.shared?.[option] ?? node?.server?.[option] ?? value
-			),
-		/** @type {Value | undefined} */ (undefined)
-	);
+	return nodes.reduce((value, node) => {
+		// TODO remove for 1.0
+		for (const thing of [node?.server, node?.shared]) {
+			if (thing && ('router' in thing || 'hydrate' in thing)) {
+				throw new Error(
+					'`export const hydrate` and `export const router` have been replaced with `export const csr`. See https://github.com/sveltejs/kit/pull/6446'
+				);
+			}
+		}
+
+		return /** @type {any} TypeScript's too dumb to understand this */ (
+			node?.shared?.[option] ?? node?.server?.[option] ?? value
+		);
+	}, /** @type {Value | undefined} */ (undefined));
 }
 
 /**